[#1] 분 단위로 데이터를 합산하는 로직 선언적으로 리펙터링
